requirements.txt
# ── Telegram бот ──
aiogram>=3.7.0
apscheduler>=3.10.0

# ── Парсинг ──
curl_cffi
pydantic
beautifulsoup4
lxml
playwright

# ── Чтение документов ──
python-docx
pypdf
striprtf
pymupdf
pytesseract
Pillow

# ── AI-анализ + Excel ──
requests
pandas
openpyxl

# ── Окружение ──
python-dotenv

# ── Тестирование ──
pytest
